requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-SocketIO
eventlet
python-dotenv
pandas
xlsxwriter
reportlab